Reduce warnings on windows.
git-svn-id: svn+ssh://edison.nersc.gov/project/projectdirs/visit/svn/visit/trunk@28216 18c085ea-50e0-402c-830e-de6fd14e8384

+++ src/avt/DBAtts/SIL/avtSIL.C
@@ -1349,7 +1349,7 @@ avtSIL::Print(ostream &out,
           case SUBSET:
           {
             out << "Set" << setTable[3*i+2] << " " << (useInfo ? perSetInfo[i].c_str() : "") << endl;
-            avtSILSet_p s = GetSILSet(i);
+            avtSILSet_p s = GetSILSet((int)i);
             s->Print(out);
             break;
           }
